+++ common/reactive/src/test/java/is/codion/common/reactive/observer/ObserverTest.java
@@ -1,0 +1,116 @@
+/*
+ * This file is part of Codion.
+ *
+ * Codion is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Codion is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Codion.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Copyright (c) 2025, Björn Darri Sigurðsson.
+ */
+package is.codion.common.reactive.observer;
+
+import is.codion.common.reactive.state.State;
+import is.codion.common.reactive.value.Value;
+import is.codion.common.reactive.value.Value.Notify;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.Objects;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+public final class ObserverTest {
+
+	@Test
+	void observeValue() {
+		AtomicInteger oneCounter = new AtomicInteger();
+		AtomicInteger twoCounter = new AtomicInteger();
+		AtomicInteger nullCounter = new AtomicInteger();
+		AtomicReference<Integer> intValue = new AtomicReference<>();
+		Value<Integer> value = Value.nullable();
+		value.when(1)
+						.run(oneCounter::incrementAndGet)
+						.when(1)
+						.accept(intValue::set)
+						.when(2)
+						.run(twoCounter::incrementAndGet)
+						.when(Objects::isNull)
+						.run(nullCounter::incrementAndGet)
+						.when(Objects::isNull)
+						.accept(intValue::set);
+
+		value.set(1);
+		assertEquals(1, intValue.get());
+		value.set(2);
+		value.clear();
+		value.set(null);// no change, no event
+		assertNull(intValue.get());
+		value.set(1);
+
+		assertEquals(1, intValue.get());
+		assertEquals(2, oneCounter.get());
+		assertEquals(1, twoCounter.get());
+		assertEquals(1, nullCounter.get());
+
+		oneCounter.set(0);
+		value = Value.builder()
+						.nonNull(0)
+						.notify(Notify.SET)
+						.build();
+		value.when(1).run(oneCounter::incrementAndGet);
+		value.set(1);
+		value.set(1);
+		assertEquals(2, oneCounter.get());
+	}
+
+	@Test
+	void observeState() {
+		AtomicInteger trueCounter = new AtomicInteger();
+		AtomicInteger falseCounter = new AtomicInteger();
+
+		State state = State.state();
+
+		state.when(true)
+						.run(trueCounter::incrementAndGet)
+						.when(false)
+						.run(falseCounter::incrementAndGet);
+
+		state.set(true);
+		state.set(true);
+		state.set(false);
+
+		assertEquals(1, trueCounter.get());
+		assertEquals(1, falseCounter.get());
+
+		state = State.builder()
+						.notify(Notify.SET)
+						.build();
+
+		trueCounter.set(0);
+		falseCounter.set(0);
+		state.when(true)
+						.run(trueCounter::incrementAndGet)
+						.when(false)
+						.run(falseCounter::incrementAndGet);
+
+		state.set(true);
+		state.set(true);
+		state.set(false);
+		state.set(false);
+
+		assertEquals(2, trueCounter.get());
+		assertEquals(2, falseCounter.get());
+	}
+}